Deduplicate `ActionMenu.tsx`
